Rename files and add license disclaimer
