+++ keras_nlp/models/__init__.py
@@ -16,10 +16,12 @@ from keras_nlp.models.bert.bert_backbone import BertBackbone
 from keras_nlp.models.bert.bert_classifier import BertClassifier
 from keras_nlp.models.bert.bert_preprocessor import BertPreprocessor
 from keras_nlp.models.bert.bert_tokenizer import BertTokenizer
-from keras_nlp.models.deberta.deberta_backbone import DebertaBackbone
-from keras_nlp.models.deberta.deberta_classifier import DebertaClassifier
-from keras_nlp.models.deberta.deberta_preprocessor import DebertaPreprocessor
-from keras_nlp.models.deberta.deberta_tokenizer import DebertaTokenizer
+from keras_nlp.models.deberta_v3.deberta_v3_backbone import DebertaBackbone
+from keras_nlp.models.deberta_v3.deberta_v3_classifier import DebertaClassifier
+from keras_nlp.models.deberta_v3.deberta_v3_preprocessor import (
+    DebertaPreprocessor,
+)
+from keras_nlp.models.deberta_v3.deberta_v3_tokenizer import DebertaTokenizer
 from keras_nlp.models.distil_bert.distil_bert_backbone import DistilBertBackbone
 from keras_nlp.models.distil_bert.distil_bert_classifier import (
     DistilBertClassifier,
