+++ keras_nlp/models/deberta_v3/deberta_v3_preprocessor_test.py
@@ -22,8 +22,10 @@ import tensorflow as tf
 from absl.testing import parameterized
 from tensorflow import keras
 
-from keras_nlp.models.deberta.deberta_preprocessor import DebertaPreprocessor
-from keras_nlp.models.deberta.deberta_tokenizer import DebertaTokenizer
+from keras_nlp.models.deberta_v3.deberta_v3_preprocessor import (
+    DebertaPreprocessor,
+)
+from keras_nlp.models.deberta_v3.deberta_v3_tokenizer import DebertaTokenizer
 
 
 class DebertaPreprocessorTest(tf.test.TestCase, parameterized.TestCase):
