+++ keras_nlp/models/deberta_v3/deberta_v3_tokenizer_test.py
@@ -22,7 +22,7 @@ import tensorflow as tf
 from absl.testing import parameterized
 from tensorflow import keras
 
-from keras_nlp.models.deberta.deberta_tokenizer import DebertaTokenizer
+from keras_nlp.models.deberta_v3.deberta_v3_tokenizer import DebertaTokenizer
 
 
 class DebertaTokenizerTest(tf.test.TestCase, parameterized.TestCase):
